+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -205,6 +205,9 @@ class Sidebar extends React.Component<any, any> {
             <Menu.Item as={NavLink} exact to="/theming" activeClassName="active">
               Theming
             </Menu.Item>
+            <Menu.Item as={NavLink} exact to="/theming-examples" activeClassName="active">
+              Theming Examples
+            </Menu.Item>
           </Menu.Menu>
         </Menu.Item>
         <Menu.Item active>